+++ src/lib.rs
@@ -42,6 +42,7 @@ pub mod network;
 pub mod new_post;
 pub mod notifications;
 pub mod parser;
+pub mod poll;
 pub mod post;
 pub mod profile;
 pub mod reply;